requirements.txt
matplotlib==2.2.3
requests==2.19.1
python-dotenv==0.9.1
progressbar2==3.38.0
